requirements.txt
SQLAlchemy==1.4.36
loguru==0.6.0
pytest==7.1.2
pytest-cov==3.0.0
email-validator==1.3.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-Migrate==3.1.0
Flask-WTF==1.0.1
Bootstrap-Flask==2.1.0
PyJWT==2.4.0
flask-mailman==0.3.0
python-dotenv==0.21.0